split server and browser files to allow optimized code
