+++ src/middlewares/index.ts
@@ -1,1 +1,1 @@
-export { default as nextI18NextMiddleware } from './next-i18next-middleware'
+export { default as nextI18NextMiddleware } from './server-next-i18next-middleware'
